% Generated by build script
\name{rw_dataset}
\alias{rw_dataset}
\title{Fetch a single Resource Watch dataset by ID}
\usage{
rw_dataset()
}
\arguments{
\item{id}{Dataset ID string}
}
\value{tibble: one row with id, name, subtitle, description, provider,}
\description{Fetch a single Resource Watch dataset by ID}
