% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_datasets}
\alias{rw_datasets}
\title{List Resource Watch datasets}
\usage{
rw_datasets(page_size = 10, page = 1)
}
\arguments{
\item{page_size}{Number of results (default 10)}

\item{page}{Page number (default 1)}
}
\value{
tibble: id, name, subtitle, provider, connectorType, published, env
}
\description{
List Resource Watch datasets
}
