% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_dataset}
\alias{rw_dataset}
\title{Fetch a single Resource Watch dataset by ID}
\usage{
rw_dataset(id)
}
\arguments{
\item{id}{Dataset ID string}
}
\value{
tibble: one row with id, name, subtitle, description, provider,
connectorType, tableName, published
}
\description{
Fetch a single Resource Watch dataset by ID
}
