% Generated by build script
\name{rw_datasets}
\alias{rw_datasets}
\title{List Resource Watch datasets}
\usage{
rw_datasets()
}
\arguments{
\item{page_size}{Number of results (default 10)}
\item{page}{Page number (default 1)}
}
\value{tibble: id, name, subtitle, provider, connectorType, published, env}
\description{List Resource Watch datasets}
